\name{call_libcoin}
\alias{call_libcoin}
\title{
    Call a libcoin C Function
}
\description{
    Calls a C function implemented in libcoin
}
\usage{
call_libcoin(routine, ...)
}
\arguments{
  \item{routine}{The name of the C function to be called.
}
  \item{\dots}{Arguments of this function.
}
}
\details{
  This function must not be used.
}
